requirements.txt
Flask
joblib
numpy
opencv-python-headless
scikit-image
dlib
pillow
pandas
requests
scikit-learn
gunicorn
